Update 정수 제곱근 판별.cpp

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.cpp
@@ -4,9 +4,10 @@
 
 using namespace std;
 
+
 long long solution(long long n) {
     long long answer = 0;
-    if(sqrt(n)-(int)sqrt(n)==0){
+    if(sqrt(n)-(long long)sqrt(n)==0){
         return (sqrt(n)+1)*(sqrt(n)+1);
     }
     else return -1;
